Add $insertBlockNode
