+++ packages/lexical-playground/src/plugins/YouTubePlugin.ts
@@ -7,17 +7,8 @@
  */
 
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
-import {
-  $createParagraphNode,
-  $getRoot,
-  $getSelection,
-  $isGridSelection,
-  $isNodeSelection,
-  $isRangeSelection,
-  COMMAND_PRIORITY_EDITOR,
-  createCommand,
-  LexicalCommand,
-} from 'lexical';
+import {$insertBlockNode} from '@lexical/utils';
+import {COMMAND_PRIORITY_EDITOR, createCommand, LexicalCommand} from 'lexical';
 import {useEffect} from 'react';
 
 import {$createYouTubeNode, YouTubeNode} from '../nodes/YouTubeNode';
@@ -35,23 +26,8 @@ export default function YouTubePlugin(): JSX.Element | null {
     return editor.registerCommand<string>(
       INSERT_YOUTUBE_COMMAND,
       (payload) => {
-        const selection = $getSelection();
         const youTubeNode = $createYouTubeNode(payload);
-        if ($isRangeSelection(selection)) {
-          const focusNode = selection.focus.getNode();
-          focusNode.getTopLevelElementOrThrow().insertAfter(youTubeNode);
-        } else if ($isNodeSelection(selection) || $isGridSelection(selection)) {
-          const nodes = selection.getNodes();
-          nodes[nodes.length - 1]
-            .getTopLevelElementOrThrow()
-            .insertAfter(youTubeNode);
-        } else {
-          const root = $getRoot();
-          root.append(youTubeNode);
-        }
-        const paragraphNode = $createParagraphNode();
-        youTubeNode.insertAfter(paragraphNode);
-        paragraphNode.select();
+        $insertBlockNode(youTubeNode);
 
         return true;
       },
